Fix memory safety issues and buffer handling bugs (#723)
- Fix memmove bug in KSFileUtils.c fillReadBuffer that used wrong byte count
- Add explicit (size_t) casts to prevent signed/unsigned conversion warnings
- Replace sprintf with snprintf to prevent potential buffer overflows
- Update test bundle accessor to use modern SPM resource bundle API

+++ Sources/KSCrashRecordingCore/KSThread.c
@@ -151,7 +151,7 @@ bool ksthread_getQueueName(const KSThread thread, char *const buffer, int bufLen
         return false;
     }
     bufLength = MIN(length, bufLength - 1);  // just strlen, without null-terminator
-    strncpy(buffer, queue_name, bufLength);
+    strncpy(buffer, queue_name, (size_t)bufLength);
     buffer[bufLength] = 0;  // terminate string
     KSLOG_TRACE("Queue label = %s", buffer);
     return true;
